refac: Split up into interface vs default impl.
