+++ optimizely/decision/engine.go
@@ -10,15 +10,3 @@ type FeatureDecision struct {
 	Type           string
 	FeatureEnabled bool
 }
-
-// DefaultDecisionEngine is used out of the box and can be replaced by the user
-type DefaultDecisionEngine struct {
-}
-
-// GetFeatureDecision returns a decision for the given feature key
-func (*DefaultDecisionEngine) GetFeatureDecision(featureKey string, userID string) FeatureDecision {
-	enabled := featureKey != "" && userID != ""
-	return FeatureDecision{
-		FeatureEnabled: enabled,
-	}
-}
